chore: Split tab view files
